Add MPI parallelism and rotatable grids to GIST (#1055)
* Use gridBin instead of Bin()

* Start the GridMover class

* Add SetTgt

* Add MoverFinish

* Add MoveGrid

* Add comment

* Add test of boxcenter keyword for grid

* Add info printout to mover. Remove mover-related code from GridAction,
use GridMover instead.

* Update dependencies

* Add missing MoverSetComm function and make void

* Clean up comments

* Add comment

* Start with lower case to reflect that function is private

* Add masterGrid_ variable to indicate which grid is being used to
determine voxels for all other grids

* Initialize grid mover for GIST

* Add mover info print

* Do mover setup

* Move the grid

* Finish moving the grid

* Add some debug info. Add ability to access max values from GridBin.

* Instead of checking cartesian deltas from the grid origin to decide if water is in a 1.5 Ang region around the grid, create a "border" grid that extends 1.5 Ang in each dimension; makes it easier to check if waters are near the grid if the grid has been rotated.

* Debug print molecule voxel # and dipole

* Dont recalculate voxel indices in searchGridNearestNeighbors6D (only
works if grid is x-aligned orthogonal); just use the already-calculated voxel indices.

* Add some more debug stuff. Use borderGrid for determining if PME solute
is on the grid

* Get rid of Gmin and Gmax

* Clean up debug info

* Remember to reset any previous rotation of the border grid

* Need to ensure the coordinates are rotated back into the reference frame
of the grid for any translation calc

* Ensure coords for calculating O-HX vecs are rotated as well

* Fix the dipole calc by rotating coords into the grid

* Clean up includes. Hide debug output behind DEBUG_GIST define

* Hide debug info

* Add some comments. Remove old code

* Update dependencies

* Add some gist file explanations

* Start MPI parallelizing gist calc

* Start syncing up more arrays

* Ensure max_nwat_ on master is updated after N_solvent_ array is synced to master

* Should be MPI_DOUBLE, not MPI_FLOAT

* Sync PME energy arrays

* Can run in parallel now

* Fix send/recv tags for mpi

* Save border grid unit cell when initialized to fix run in parallel

* Increase test tolerance for MPI

* Remove old code

* Past in const ref instead of a copy

* Simplify setting different tolerances based on the binary type

* Add openmp key

* Use processes

* Start NDiff

* Finish up initial ndiff

* Enable --ndiff flag. Pass in tolerance value as a string initially

* Change so cpptraj ndiff accepts arguments the same way ndiff.awk does

* Dont assume CPPTRAJ_NDIFF is awk

* Use ndiff built into cpptraj

* Accept -v ABSERR=<tol> at the same time as -v RELERR=<tol>

* Rework ndiff so error message only printed if an error actually occurs; also makes it so the file is short message actually shows up.

* Look at exit status of CPPTRAJ_NDIFF and CPPTRAJ_DIFF instead of size of temp file; avoids false positives in cases where extraneous error messages get printed out.

* Add test for ndiff functionality

* Enable ndiff test

* Add print subtimers

* Move timer stop

* Split off translational entropy calc in preparation for parallelizing it

* Initial parallel trans entropy calc for gist

* Enable parallel post calc

* Make routine const

* Make it so the calc can be divided

* Divide grid points

* Finish calculation in parallel; brute force requiring all voxels to be
synced up on all processes

* Add timer

* Hide some debug info

* Ensure border grid is kept centered with regular grid

* Ensure on grid water-in-voxel coords are properly translated w.r.t. the
grid center to ensure the translational entropy calculation works
properly

* Add order calc debug info

* Save nearest water indices for debug

* Add imaging to the order parameter calc

* Put getting rid of multiples of box lengths in the loop

* Add NeedsMove()

* Set rotation matrix default to no rotation (identity). Use NoMove
instead of RotationHappened

* Use NeedsMove

* Fix comment

* Remove the doRotate var, no longer needed

* Start pairlist version of order

* Save which cell each atom is in

* Finish pairlist order calc

* Fix debug statement

* Add distance squared to debug output. Put trans entropy debug behind a
different ifdef

* Need to do the pairlist loop a certain way since it expects all
interactions in the forward direction

* Add the order calc. Ensure it happens in the same order as the original
PL calc (i.e. vectors are in the same orientation).

* Add way more code doc so I dont forget how to properly use PairList

* Comment out cell of index

* Increase the cutoff, sometimes missed the 4th farthest water with 8 Ang.
Improve debug output

* Hide some debug info

* Remove unused code

* Simplify the routine

* Add order pairlist cutoff arg

* Add nopl and plcut keywords

* Protect against segfault

* Hide debug message

* Add name and rmsfit keywords

* Should be processes

* Protect against too many procs

* Protect against too many procs

* Protect in parallel. Set parallel tolerance

* Add entropy communication timer. Add trans entropy calc timer to total
in MPI since it occurs outside the print function. Make entropy timers
relative to total for same reason.

* Fix clean rules

* Hide some debug info

* Update dependencies

* Remove unused code

* Document tag for diffusion parallel multiple time origin

* Add a note about GIST entropy MPI parallelism

* 6.22.0. Minor version bump for addition of MPI parallel to GIST,
  addition of 'rmsfit' keyword to GIST, addition of imaging to GIST
order calculation (and pairlist for speedup), and addition of fast
'ndiff' functionality to cpptraj.

* When updating relative cpptraj binary test path, update ndiff path as
well.

---------

+++ src/DataSet_GridDbl.cpp
@@ -27,10 +27,10 @@ int DataSet_GridDbl::SyncGrid(size_t total, std::vector<int> const& rank_frames,
 {
   if (commIn.Master()) {
     std::vector<double> buf( grid_.size() );
-    commIn.ReduceMaster( &(buf[0]), &(grid_[0]), grid_.size(), MPI_FLOAT, MPI_SUM );
+    commIn.ReduceMaster( &(buf[0]), &(grid_[0]), grid_.size(), MPI_DOUBLE, MPI_SUM );
     std::copy( buf.begin(), buf.end(), grid_.begin() );
   } else
-    commIn.ReduceMaster( 0,         &(grid_[0]), grid_.size(), MPI_FLOAT, MPI_SUM );
+    commIn.ReduceMaster( 0,         &(grid_[0]), grid_.size(), MPI_DOUBLE, MPI_SUM );
   return 0;
 }
 #endif
